+++ pkgs/test/lib/src/runner/node/socket_channel.dart
@@ -1,46 +1,41 @@
 // Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-
-@JS()
-library;
-
 import 'dart:async';
 import 'dart:convert';
+import 'dart:js_interop';
 
-import 'package:js/js.dart';
 import 'package:stream_channel/stream_channel.dart';
 
-@JS('require')
-external _Net _require(String module);
-
 @JS('process.argv')
-external List<String> get _args;
+external JSArray<JSString> get _args;
 
-@JS()
-class _Net {
+extension type _Net._(JSObject _) {
   external _Socket connect(int port);
 }
 
-@JS()
-class _Socket {
-  external void setEncoding(String encoding);
-  external void on(String event, void Function(String chunk) callback);
-  external void write(String data);
+extension type _Socket._(JSObject _) {
+  external void setEncoding(JSString encoding);
+  external void on(JSString event, JSFunction callback);
+  external void write(JSString data);
 }
 
 /// Returns a [StreamChannel] of JSON-encodable objects that communicates over a
 /// socket whose port is given by `process.argv[2]`.
-StreamChannel<Object?> socketChannel() {
-  var net = _require('net');
-  var socket = net.connect(int.parse(_args[2]));
-  socket.setEncoding('utf8');
+Future<StreamChannel<Object?>> socketChannel() async {
+  final net = (await importModule('node:net'.toJS).toDart) as _Net;
+
+  var socket = net.connect(int.parse(_args.toDart[2].toDart));
+  socket.setEncoding('utf8'.toJS);
 
   var socketSink = StreamController<Object?>(sync: true)
-    ..stream.listen((event) => socket.write('${jsonEncode(event)}\n'));
+    ..stream.listen((event) => socket.write('${jsonEncode(event)}\n'.toJS));
 
   var socketStream = StreamController<String>(sync: true);
-  socket.on('data', allowInterop(socketStream.add));
+  socket.on(
+    'data'.toJS,
+    ((JSString chunk) => socketStream.add(chunk.toDart)).toJS,
+  );
 
   return StreamChannel.withCloseGuarantee(
       socketStream.stream.transform(const LineSplitter()).map(jsonDecode),