+++ pkgs/test_core/lib/src/runner/wasm_compiler_pool.dart
@@ -17,8 +17,13 @@ import 'suite.dart';
 ///
 /// This limits the number of compiler instances running concurrently.
 class WasmCompilerPool extends CompilerPool {
+  /// Extra arguments to pass to `dart compile js`.
+  final List<String> _extraArgs;
+
   /// The currently-active dart2wasm processes.
   final _processes = <Process>{};
+
+  WasmCompilerPool([this._extraArgs = const []]);
 
   /// Compiles [code] to [path].
   ///
@@ -41,6 +46,7 @@ class WasmCompilerPool extends CompilerPool {
         for (var experiment in enabledExperiments)
           '--enable-experiment=$experiment',
         '-O0',
+        ..._extraArgs,
         '-o',
         outWasmPath,
         wrapperPath,